Add vitest testing and parseResponse tests
